Convert TestEnvironmentIntegrationTest to new junit multiversion test
