--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsReferencingSlf4jWithModularJava/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'modular'
