--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsReferencingSlf4jWithModularJava/src/test/java/module-info.java
@@ -1,5 +1,0 @@
-module org.gradle.example {
-    exports org.gradle.example;
-    requires junit;
-    requires org.slf4j;
-}
