--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSecurityManager/build.gradle
@@ -1,19 +1,0 @@
-apply plugin: "java"
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(11)
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation "junit:junit:4.13"
-}
-
-test {
-    systemProperties 'java.security.manager': 'org.gradle.MySecurityManager'
-}
