--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSecurityManager/src/test/java/org/gradle/JUnitTest.java
@@ -1,14 +1,0 @@
-package org.gradle;
-
-import org.junit.Test;
-
-import static junit.framework.Assert.*;
-import static org.junit.Assert.assertEquals;
-
-public class JUnitTest {
-    @Test
-    public void mySecurityManagerIsUsed() throws ClassNotFoundException {
-        assertTrue(System.getSecurityManager() instanceof MySecurityManager);
-        assertEquals(ClassLoader.getSystemClassLoader(), MySecurityManager.class.getClassLoader());
-    }
-}
