--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSecurityManager/src/test/java/org/gradle/MySecurityManager.java
@@ -1,15 +1,0 @@
-package org.gradle;
-
-import java.security.Permission;
-
-import static org.junit.Assert.assertEquals;
-
-public class MySecurityManager extends SecurityManager {
-    public MySecurityManager() {
-        assertEquals(getClass().getName(), System.getProperty("java.security.manager"));
-    }
-
-    @Override
-    public void checkPermission(Permission permission) {
-    }
-}
