--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSystemClassLoader/build.gradle
@@ -1,13 +1,0 @@
-apply plugin: "java"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation "junit:junit:4.13"
-}
-
-test {
-    systemProperties 'java.system.class.loader':'org.gradle.MySystemClassLoader'
-}
