--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSystemClassLoader/src/test/java/org/gradle/JUnitTest.java
@@ -1,13 +1,0 @@
-package org.gradle;
-
-import org.junit.Test;
-
-import static junit.framework.Assert.*;
-
-public class JUnitTest {
-    @Test
-    public void mySystemClassLoaderIsUsed() throws ClassNotFoundException {
-        assertTrue(ClassLoader.getSystemClassLoader() instanceof MySystemClassLoader);
-        assertEquals(getClass().getClassLoader(), ClassLoader.getSystemClassLoader().getParent());
-    }
-}
