--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSystemClassLoaderAndJavaAgent/build.gradle
@@ -1,21 +1,0 @@
-apply plugin: "java"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation "junit:junit:4.13"
-}
-
-jar {
-    manifest {
-        attributes 'Premain-Class': 'org.gradle.MyAgent'
-    }
-}
-
-test {
-    dependsOn jar
-    systemProperties 'java.system.class.loader':'org.gradle.MySystemClassLoader'
-    jvmArgs "-javaagent:${jar.archiveFile.asFile.get()}"
-}
