--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/TestEnvironmentIntegrationTest/canRunTestsWithCustomSystemClassLoaderAndJavaAgent/src/test/java/org/gradle/JUnitTest.java
@@ -1,17 +1,0 @@
-package org.gradle;
-
-import org.junit.Test;
-
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertTrue;
-
-public class JUnitTest {
-    @Test
-    public void mySystemClassLoaderIsUsed() throws ClassNotFoundException {
-        assertEquals("true", System.getProperty("using.custom.agent"));
-
-        // This test class should be loaded via the custom system ClassLoader
-        assertTrue(ClassLoader.getSystemClassLoader() instanceof MySystemClassLoader);
-        assertTrue(getClass().getClassLoader() == ClassLoader.getSystemClassLoader());
-    }
-}
